+++ lib/legion/extensions/swarm_github.rb
@@ -10,6 +10,8 @@ require 'legion/extensions/swarm_github/helpers/diff_chunker'
 require 'legion/extensions/swarm_github/runners/review_poster'
 require 'legion/extensions/swarm_github/runners/review_notifier'
 require 'legion/extensions/swarm_github/runners/pr_pipeline'
+require 'legion/extensions/swarm_github/runners/extension_lifecycle'
+require 'legion/extensions/swarm_github/actors/lifecycle_subscriber' if defined?(Legion::Extensions::Actors::Subscription)
 
 module Legion
   module Extensions